use int64_t instead of int

+++ mjolnir/core/UnlimitedGridCellList.hpp
@@ -280,9 +280,9 @@ inline std::size_t
 UnlimitedGridCellList<traitsT, N>::calc_index(
         const coordinate_type& pos) const noexcept
 {
-    const auto x = static_cast<int>(std::floor(pos[0]*r_cell_size_)) % dim;
-    const auto y = static_cast<int>(std::floor(pos[1]*r_cell_size_)) % dim;
-    const auto z = static_cast<int>(std::floor(pos[2]*r_cell_size_)) % dim;
+    const auto x = static_cast<std::int64_t>(std::floor(pos[0]*r_cell_size_)) % dim;
+    const auto y = static_cast<std::int64_t>(std::floor(pos[1]*r_cell_size_)) % dim;
+    const auto z = static_cast<std::int64_t>(std::floor(pos[2]*r_cell_size_)) % dim;
     return calc_index((x<0) ? x+dim : x, (y<0) ? y+dim : y, (z<0) ? z+dim : z);
 }
